+++ packages/core/src/_test/utils.ts
@@ -312,15 +312,3 @@ export async function getMetrics(port: number) {
   const response = await fetch(`http://localhost:${port}/metrics`);
   return await response.text();
 }
-
-export async function drainAsyncGenerator<t extends unknown[]>(
-  asyncGenerator: AsyncGenerator<t>,
-) {
-  const result = [] as unknown as t;
-
-  for await (const x of asyncGenerator) {
-    result.push(...x);
-  }
-
-  return result;
-}
