+++ packages/core/src/sync-historical/service.test.ts
@@ -3,12 +3,9 @@ import {
   setupDatabaseServices,
   setupIsolatedDatabase,
 } from "@/_test/setup.js";
-import {
-  drainAsyncGenerator,
-  getEventsErc20,
-  publicClient,
-} from "@/_test/utils.js";
+import { getEventsErc20, publicClient } from "@/_test/utils.js";
 import { maxCheckpoint, zeroCheckpoint } from "@/utils/checkpoint.js";
+import { drainAsyncGenerator } from "@/utils/drainAsyncGenerator.js";
 import { toLowerCase } from "@/utils/lowercase.js";
 import { wait } from "@/utils/wait.js";
 import { beforeEach, expect, test, vi } from "vitest";
